code refactoring
include cccv variable in class
clean index after check
