+++ indexer/check.go
@@ -3,54 +3,26 @@ package indexer
 import (
 	"fmt"
 	"io"
-	"time"
 
 	"github.com/aergoio/aergo-indexer-2.0/indexer/db"
 	doc "github.com/aergoio/aergo-indexer-2.0/indexer/documents"
 )
 
 // Start setups the indexer
-func (ns *Indexer) RunCheckIndex(startFrom uint64, stopAt uint64) error {
+func (ns *Indexer) RunCheckIndex(startFrom uint64, stopAt uint64) {
 	fmt.Println("=======> Start Check index ..")
 
-	aliasName := ns.aliasNamePrefix + "block"
-	for {
-		_, _, err := ns.db.GetExistingIndexPrefix(aliasName, "block")
-
-		if err == nil {
-			break
-		}
-		time.Sleep(10 * time.Second)
-	}
-
-	ns.CreateIndexIfNotExists("block")
-	ns.CreateIndexIfNotExists("tx")
-	ns.CreateIndexIfNotExists("name")
-	ns.CreateIndexIfNotExists("token")
-	ns.CreateIndexIfNotExists("contract")
-	ns.CreateIndexIfNotExists("token_transfer")
-	ns.CreateIndexIfNotExists("account_tokens")
-	ns.CreateIndexIfNotExists("nft")
-	ns.CreateIndexIfNotExists("account_balance")
-
-	// init config
-	ns.init_cccv_nft()
 	if stopAt == 0 {
-		ns.lastHeight = ns.GetBestBlockFromClient() - 1
-	} else {
-		ns.lastHeight = stopAt
+		stopAt = ns.GetBestBlockFromClient() - 1
 	}
-	ns.bulkSize = 4000
-	ns.batchTime = 60 * time.Second
-	ns.minerNum = 32
-	ns.grpcNum = 16
-
-	ns.fixIndex(startFrom, ns.lastHeight)
-	return nil
+	ns.fixIndex(startFrom, stopAt)
+	if ns.cleanMode {
+		ns.cleanIndex()
+	}
 }
 
-func (ns *Indexer) fixIndex(Start_Pos uint64, End_Pos uint64) {
-	ns.log.Info().Uint64("startFrom", Start_Pos).Uint64("stopAt", End_Pos).Msg("Check Block range")
+func (ns *Indexer) fixIndex(startFrom uint64, stopAt uint64) {
+	ns.log.Info().Uint64("startFrom", startFrom).Uint64("stopAt", stopAt).Msg("Check Block range")
 	ns.StartBulkChannel()
 
 	var block doc.DocType
@@ -63,8 +35,8 @@ func (ns *Indexer) fixIndex(Start_Pos uint64, End_Pos uint64) {
 		Size:         10000,
 		SortField:    "no",
 		SortAsc:      false,
-		From:         int(Start_Pos),
-		To:           int(End_Pos),
+		From:         int(startFrom),
+		To:           int(stopAt),
 	}, func() doc.DocType {
 		block := new(doc.EsBlock)
 		block.BaseEsType = new(doc.BaseEsType)
@@ -72,9 +44,9 @@ func (ns *Indexer) fixIndex(Start_Pos uint64, End_Pos uint64) {
 		return block
 	})
 
-	prevBlockNo := End_Pos + 1
+	prevBlockNo := stopAt + 1
 	missingBlocks := uint64(0)
-	blockNo := Start_Pos + 1
+	blockNo := startFrom + 1
 	for {
 		block, err = scroll.Next()
 		if err == io.EOF {
@@ -92,7 +64,7 @@ func (ns *Indexer) fixIndex(Start_Pos uint64, End_Pos uint64) {
 		if blockNo >= prevBlockNo {
 			continue
 		}
-		if blockNo <= Start_Pos {
+		if blockNo <= startFrom {
 			break
 		}
 		if blockNo < prevBlockNo-1 {
@@ -102,11 +74,152 @@ func (ns *Indexer) fixIndex(Start_Pos uint64, End_Pos uint64) {
 		prevBlockNo = blockNo
 	}
 
-	if blockNo != Start_Pos && prevBlockNo > Start_Pos {
-		missingBlocks = missingBlocks + (prevBlockNo - Start_Pos)
-		ns.InsertBlocksInRange(Start_Pos, prevBlockNo-1)
+	if blockNo != startFrom && prevBlockNo > startFrom {
+		missingBlocks = missingBlocks + (prevBlockNo - startFrom)
+		ns.InsertBlocksInRange(startFrom, prevBlockNo-1)
 	}
 
 	ns.StopBulkChannel()
 	ns.log.Info().Uint64("missing", missingBlocks).Msg("Done with consistency check")
+}
+
+// Start clean the indexer
+func (ns *Indexer) cleanIndex() error {
+	fmt.Println("=======> Start Clean index ..")
+
+	// 1. get token list
+	scrollToken := ns.db.Scroll(db.QueryParams{
+		IndexName:    ns.indexNamePrefix + "token",
+		TypeName:     "_doc",
+		Size:         10000,
+		SelectFields: []string{"name"},
+		SortField:    "blockno",
+		SortAsc:      true,
+	}, func() doc.DocType {
+		token := new(doc.EsToken)
+		token.BaseEsType = new(doc.BaseEsType)
+		return token
+	})
+
+	tokens := make(map[string]bool)
+	for {
+		token, err := scrollToken.Next()
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			ns.log.Warn().Err(err).Msg("Failed to get token")
+			return err
+		}
+		tokens[token.(*doc.EsToken).Name] = true
+	}
+
+	// 2. get token transfer
+	scrollTransfer := ns.db.Scroll(db.QueryParams{
+		IndexName:    ns.indexNamePrefix + "token_transfer",
+		Size:         10000,
+		SelectFields: []string{"address"},
+		SortField:    "blockno",
+		SortAsc:      true,
+	}, func() doc.DocType {
+		transfer := new(doc.EsTokenTransfer)
+		transfer.BaseEsType = new(doc.BaseEsType)
+		return transfer
+	})
+
+	for {
+		transfer, err := scrollTransfer.Next()
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			ns.log.Warn().Err(err).Msg("Failed to get token transfer")
+			return err
+		}
+		if _, ok := tokens[transfer.(*doc.EsTokenTransfer).TokenAddress]; !ok {
+			ns.log.Info().Str("token", transfer.(*doc.EsTokenTransfer).TokenAddress).Msg("Delete token transfer")
+			_, err := ns.db.Delete(db.QueryParams{
+				IndexName: ns.indexNamePrefix + "token_transfer",
+				StringMatch: &db.StringMatchQuery{
+					Field: "address",
+					Value: transfer.(*doc.EsTokenTransfer).TokenAddress,
+				},
+			})
+			if err != nil {
+				ns.log.Warn().Err(err).Msg("Failed to delete token transfer")
+				return err
+			}
+		}
+	}
+
+	// 3. get account_tokens
+	scrollAccount := ns.db.Scroll(db.QueryParams{
+		IndexName:    ns.indexNamePrefix + "account_tokens",
+		Size:         10000,
+		SelectFields: []string{"address"},
+		SortField:    "ts",
+		SortAsc:      true,
+	}, func() doc.DocType {
+		account := new(doc.EsAccountTokens)
+		account.BaseEsType = new(doc.BaseEsType)
+		return account
+	})
+
+	for {
+		account, err := scrollAccount.Next()
+		if err == io.EOF {
+			break
+		} else if err != nil {
+			ns.log.Warn().Err(err).Msg("Failed to get account token")
+			return err
+		}
+		if _, ok := tokens[account.(*doc.EsAccountTokens).TokenAddress]; !ok {
+			ns.log.Info().Str("token", account.(*doc.EsAccountTokens).TokenAddress).Msg("Delete account token")
+			ns.db.Delete(db.QueryParams{
+				IndexName: ns.indexNamePrefix + "account_tokens",
+				StringMatch: &db.StringMatchQuery{
+					Field: "address",
+					Value: account.(*doc.EsAccountTokens).TokenAddress,
+				},
+			})
+		}
+	}
+
+	// 4. get account_balance
+	scrollBalance := ns.db.Scroll(db.QueryParams{
+		IndexName:    ns.indexNamePrefix + "account_balance",
+		Size:         10000,
+		SelectFields: []string{"_id"},
+		SortField:    "blockno",
+		SortAsc:      true,
+	}, func() doc.DocType {
+		balance := new(doc.EsAccountBalance)
+		balance.BaseEsType = new(doc.BaseEsType)
+		return balance
+	})
+
+	for {
+		document, err := scrollBalance.Next()
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			ns.log.Warn().Err(err).Msg("Failed to get account balance")
+			return err
+		}
+		balance := document.(*doc.EsAccountBalance)
+
+		// delete alias account balance only
+		if doc.IsBalanceNotResolved(balance.Id) == true {
+			ns.log.Info().Str("id", balance.Id).Msg("Delete account balance")
+			ns.db.Delete(db.QueryParams{
+				IndexName: ns.indexNamePrefix + "account_balance",
+				StringMatch: &db.StringMatchQuery{
+					Field: "_id",
+					Value: balance.Id,
+				},
+			})
+		}
+	}
+	return nil
 }
